Improve identifiers, mostly eliminating Guids
Guids are convenient but incredibly ugly and hard to dictate and
should never be displayed anywhere in the system's UI (in notifications, urls,
etc.). Yes, URIs are UIs too, so, the fewer Guids, the better.

This improves things in a few ways:
1. DomainEvents used Guids always, but there was already an equivalent uniqueness
   guarantee from the StreamStone's usage of PartitionKey (PersonId) plus an incrementing
   value (stream/object Version), padded with '0' to 10 chars. In addition to adding
   consistency, adopting a similar scheme for the DomainEvent.EventId also allows us to
   quickly determine the domain object that originally triggered that event, even when
   scanning the Event table.
2. RequestId also used a Guid, but since requests are always created by a person, we can
   use that as a prefix and that already gives us significant uniqueness scope reduction.
   In this case though, instead of depending on the source person incrementing version,
   it was useful to have a more "displayable" identifier, since the request will eventually
   become something with some sort of permalink which can be shared even.

   So we leverage the existing Base62 encoding to create a short string from a number. This
   number, to guarantee uniqueness, takes the total seconds elapsed since quarantine started
   in Argentina (some "magic number") and divides that by the person version, which is guaranteed
   to be always incrementing. Since both values are guaranteed to be always incrementing, we
   can guarantee the resulting hash will also be unique (within a person's requests).

   And since now the PersonId is directly encoded in the RequestId, we also eliminated
   the additional property from serialization in a few places too.

3. Since we started using the Base62 encoder now, we also bring back the two functions,
   `/encoder/encode?value` and `/encoder/decode?value` for easy testing/troubleshooting.

4. Minor renaming for consistency from Person.Id > Person.PersonId, which is used in
   events and other types.

The new IIdentifiable interface is optional for `DomainObject`-derived classes, but
is what enables the improved event ids when implemented (we do in both Person and Request,
obviously).

+++ Core/Infrastructure/Base62.cs
@@ -12,7 +12,7 @@ public static class Base62
     /// <summary>
     /// Encodes a numeric value into a base62 string.
     /// </summary>
-    public static string Encode(int value)
+    public static string Encode(long value)
     {
         // TODO: I'm almost sure there's a more succint way 
         // of doing this with LINQ and Aggregate, but just 
@@ -31,18 +31,18 @@ public static class Base62
     /// <summary>
     /// Decodes a base62 string into its original numeric value.
     /// </summary>
-    public static int Decode(string value)
-        => value.Aggregate(0, (current, c) => current * 62 + FromBase62(c));
+    public static long Decode(string value)
+        => value.Aggregate(0L, (current, c) => current * 62 + FromBase62(c));
 
-    static char ToBase62(int d) => d switch
+    static char ToBase62(long d) => d switch
     {
-        int v when v < 10 => (char)('0' + d),
-        int v when v < 36 => (char)('A' + d - 10),
-        int v when v < 62 => (char)('a' + d - 36),
+        long v when v < 10 => (char)('0' + d),
+        long v when v < 36 => (char)('A' + d - 10),
+        long v when v < 62 => (char)('a' + d - 36),
         _ => throw new ArgumentException($"Cannot encode digit {d} to base 62.", nameof(d)),
     };
 
-    static int FromBase62(char c) => c switch
+    static long FromBase62(char c) => c switch
     {
         char v when c >= 'a' && v <= 'z' => 36 + c - 'a',
         char v when c >= 'A' && v <= 'Z' => 10 + c - 'A',
